+++ ChangeLog.md
@@ -1,4 +1,5 @@
 # Pending
+* System interface for expandability
 * EXP: Generate normal/disadvantage results in Spot effect for all Perception rolls
 * EXP: Implement Dim/Dark flag effects on Perception checks for Foundry darkvision model
